+++ src/utils/icons/index.js
@@ -1,0 +1,17 @@
+import add from './add'
+import createElement from './createElement'
+import getSymbol from './getSymbol'
+import getSymbols from './getSymbols'
+import icon from './icon'
+import paint from './paint'
+import remove from './remove'
+
+export default {
+  add,
+  createElement,
+  getSymbol,
+  getSymbols,
+  icon,
+  paint,
+  remove
+}