+++ src/utils/icons/symbols.js
@@ -1,0 +1,5 @@
+import DEFAULTS from './defaults'
+
+const SYMBOLS = [...DEFAULTS]
+
+export default SYMBOLS